refactoring - file tree, namings etc.
