+++ .github/workflows/ci-matrix-release.yml
@@ -20,8 +20,6 @@ permissions:
 jobs:
   tests:
     uses: ./.github/workflows/tests.yml
-    secrets:
-      SAFETY_API_KEY: ${{ secrets.SAFETY_API_KEY }}
   
   update-badges:
     if: github.ref == 'refs/heads/main' || startsWith(github.ref, 'refs/tags/')
