+++ main.py
@@ -1,10 +1,13 @@
 import sys
+# import os
+# sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '../..')))
 from PyQt6.QtWidgets import QApplication
 from config import SETTINGS_PATH
-from sponsor.settings_manager import SettingsManager
-from sponsor.init_logger import init_logger
-from ui.main_ui import TSAController
+from models.settings_model import SettingsManager
+from utils.init_logger import init_logger
+from controllers.main_controller import TSAController
 
+# print(sys.path)
 
 def main():
     settings_manager = SettingsManager(SETTINGS_PATH)
