+++ views/menu_view.py
