+++ views/navigation_view.py
